--- src/main/java/alz/board/domain/Criteria.java
@@ -1,22 +1,0 @@
-package alz.board.domain;
-
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Setter
-@ToString
-public class Criteria {
-
-	private int pageNum;
-	private int amount;
-	
-	public Criteria() {
-		this(1,10);
-	}
-	public Criteria(int pageNum, int amount) {
-		this.pageNum = pageNum;
-		this.amount = amount;
-	}
-}
