+++ src/main/java/alz/board/mapper/BoardMapper.java
@@ -3,7 +3,6 @@ package alz.board.mapper;
 import java.util.List;
 import alz.board.domain.BoardCriteria;
 import alz.board.domain.BoardDTO;
-import alz.board.domain.Criteria;
 
 public interface BoardMapper {
 	
@@ -14,8 +13,6 @@ public interface BoardMapper {
 	List<BoardDTO> selectAll();
 
 	List<BoardDTO> selectWithPaging(BoardCriteria cri);
-	
-	List<BoardDTO> getListWithPaging(Criteria cri);
 	
 	int updateById(BoardDTO board);
