+++ src/main/java/alz/board/service/BoardService.java
@@ -4,7 +4,6 @@ import java.util.List;
 
 import alz.board.domain.BoardCriteria;
 import alz.board.domain.BoardDTO;
-import alz.board.domain.Criteria;
 
 public interface BoardService {
